- Added MergeTest unit test (fixes https://github.com/belaban/jgroups-raft/issues/125) - MergeTest checks for more than 1 leader in the same term (violation of spec)

+++ src/org/jgroups/protocols/raft/VoteRequest.java
@@ -11,7 +11,7 @@ import java.util.function.Supplier;
 public class VoteRequest extends RaftHeader {
 
     public VoteRequest() {}
-    public VoteRequest(int term) {
+    public VoteRequest(long term) {
         super(term);
     }
